Restructure repo to golang-standards/project-layout
Move all Go packages under internal/ (application-private), the binary
entry point under cmd/eth2-monitor/, and rename the module from the bare
'eth2-monitor' to github.com/stakefish/eth2-monitor:

  main.go               -> cmd/eth2-monitor/main.go
  cmd/root.go           -> internal/cli/root.go      (package cmd -> cli)
  cmd/opts/             -> internal/opts/
  beaconchain/          -> internal/beaconchain/
  spec/                 -> internal/spec/
  pkg/                  -> internal/monitoring/      (package pkg -> monitoring)

Updates Makefile ldflags target to the new cli package path and switches
the build/e2e-test paths to ./cmd/eth2-monitor and
./internal/beaconchain/.... Dockerfile build target gets the explicit
./cmd/eth2-monitor package argument now that main.go is no longer at the
repo root. Dockerfile, test-env/, and docs/ stay at the root - the
project-layout's build/package and deployments slots are not worth the
churn here.

Verified via: go build ./..., go vet ./..., make test (all pass),
make build then bin/eth2-monitor version (prints git-derived version,
proving the new -X .../internal/cli.version=... ldflag landed).
