+++ Makefile
@@ -1,6 +1,6 @@
 VERSION = $(shell git describe --tags --abbrev=8 2>/dev/null)
 
-LDFLAGS += -X eth2-monitor/cmd.version=${VERSION}
+LDFLAGS += -X github.com/stakefish/eth2-monitor/internal/cli.version=${VERSION}
 
 .PHONY: all build eth2-monitor lint test test-e2e
 all: build
@@ -10,7 +10,7 @@ build: eth2-monitor
 eth2-monitor:
 	-@mkdir -p bin
 	-@rm -f bin/$@
-	go build -ldflags '$(LDFLAGS)' -o bin/$@ .
+	go build -ldflags '$(LDFLAGS)' -o bin/$@ ./cmd/eth2-monitor
 
 lint:
 	golangci-lint run ./...
@@ -19,4 +19,4 @@ test:
 	go test -cover ./...
 
 test-e2e:
-	go test -tags=e2e -count=1 -timeout=2m ./beaconchain/...
+	go test -tags=e2e -count=1 -timeout=2m ./internal/beaconchain/...
