+++ internal/beaconchain/caplin_compat.go
