+++ internal/beaconchain/metrics.go
