+++ internal/monitoring/cache_test.go
@@ -1,4 +1,4 @@
-package pkg
+package monitoring
 
 import (
 	"os"
