+++ internal/monitoring/epoch_context.go
@@ -1,4 +1,4 @@
-package pkg
+package monitoring
 
 import (
 	"context"
@@ -6,8 +6,8 @@ import (
 	"slices"
 	"time"
 
-	"eth2-monitor/beaconchain"
-	"eth2-monitor/spec"
+	"github.com/stakefish/eth2-monitor/internal/beaconchain"
+	"github.com/stakefish/eth2-monitor/internal/spec"
 
 	v1 "github.com/attestantio/go-eth2-client/api/v1"
 	"github.com/attestantio/go-eth2-client/spec/electra"
