+++ internal/monitoring/metrics_test.go
@@ -1,4 +1,4 @@
-package pkg
+package monitoring
 
 import (
 	"testing"
