+++ internal/monitoring/proposals_test.go
@@ -1,4 +1,4 @@
-package pkg
+package monitoring
 
 import (
 	"strings"
