+++ internal/monitoring/set.go
@@ -1,4 +1,4 @@
-package pkg
+package monitoring
 
 import (
 	"fmt"
