+++ internal/monitoring/set_test.go
@@ -1,4 +1,4 @@
-package pkg
+package monitoring
 
 import (
 	"slices"
