+++ internal/spec/consts.go
